% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localkernel_utils.R
\name{get_kernel_vals}
\alias{get_kernel_vals}
\title{Compute unnormalized kernel values.}
\usage{
get_kernel_vals(x, center, bw, kernel)
}
\arguments{
\item{x}{Numeric vector.}

\item{center}{Numeric value to center the kernel.}

\item{bw}{Bandwidth of the kernel.}

\item{kernel}{One of \code{c("gaussian", "uniform", "triangular")}.}
}
\value{
Kernel values for \code{x} centered at \code{center}.
}
\description{
Helper function used internally by \code{mice.impute.sl}, adapted
from \code{superMICE::gaussianKernel}, \code{superMICE::uniformKernel}, and
\code{superMICE::triangularKernel} by Laqueur et al. 2022, see
\url{https://github.com/abshev/superMICE}.
}
\references{
Laqueur HS, Shev AB, Kagawa RMC. SuperMICE: An Ensemble Machine Learning Approach to Multiple Imputation by Chained Equations. American Journal of Epidemiology. 2022;191(3):516-525.
}
\keyword{internal}
