% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localkernel_utils.R
\name{jackknife_var}
\alias{jackknife_var}
\title{Estimate the jackknife variance from a matrix of kernel values.}
\usage{
jackknife_var(j, kernel_mat, delta, y)
}
\arguments{
\item{j}{Index of the deleted observation in the jackknife procedure.}

\item{kernel_mat}{Matrix of kernel values centered at \code{j} of
shape \code{(n-1)} by \code{k}, where \code{n} is the total number of
observations and \code{k} is the number of candidate bandwidths.}

\item{delta}{Missingness indicator of \code{y}.}

\item{y}{Variable to be imputed.}
}
\value{
Numeric variance estimate.
}
\description{
Helper function used internally by \code{mice.impute.sl}, adapted
from \code{superMICE::jackknifeVariance} by Laqueur et al. 2022, see
\url{https://github.com/abshev/superMICE}.
}
\references{
Laqueur HS, Shev AB, Kagawa RMC. SuperMICE: An Ensemble Machine Learning Approach to Multiple Imputation by Chained Equations. American Journal of Epidemiology. 2022;191(3):516-525.
}
\keyword{internal}
