% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localkernel_utils.R
\name{local_imputation}
\alias{local_imputation}
\title{Impute values based on a local normal distribution with a
kernel-weighted variance estimate.}
\usage{
local_imputation(i, preds, y, delta, bw, kernel)
}
\arguments{
\item{i}{Index of the missing value to be imputed.}

\item{preds}{Predictions for missing values.}

\item{y}{Variable to be imputed.}

\item{delta}{Missingness indicator of \code{y}.}

\item{bw}{Numeric vector of kernel bandwidths.}

\item{kernel}{One of \code{c("gaussian", "uniform", "triangular")}.}
}
\value{
Numeric vector of imputed values.
}
\description{
Helper function used internally by \code{mice.impute.sl}, adapted
from \code{superMICE::localImputation} by Laqueur et al. 2022, see
\url{https://github.com/abshev/superMICE}.
}
\references{
Laqueur HS, Shev AB, Kagawa RMC. SuperMICE: An Ensemble Machine Learning Approach to Multiple Imputation by Chained Equations. American Journal of Epidemiology. 2022;191(3):516-525.
}
\keyword{internal}
