% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localkernel_utils.R
\name{select_bandwidth}
\alias{select_bandwidth}
\title{Select an optimal bandwidth among a set of candidates using a
jackknife procedure.}
\usage{
select_bandwidth(i, bw_grid, preds, y, delta, kernel)
}
\arguments{
\item{i}{Index of the missing value to be imputed.}

\item{bw_grid}{Candidate bandwidth values.}

\item{preds}{Predictions for missing values.}

\item{y}{Variable to be imputed.}

\item{delta}{Missingness indicator of \code{y}.}

\item{kernel}{One of \code{c("gaussian", "uniform", "triangular")}.}
}
\value{
Optimal bandwidth value.
}
\description{
Helper function used internally by \code{mice.impute.sl}, adapted
from \code{superMICE::jackknifeBandwidthSelection} by Laqueur et al. 2022,
see \url{https://github.com/abshev/superMICE}.
}
\references{
Laqueur HS, Shev AB, Kagawa RMC. SuperMICE: An Ensemble Machine Learning Approach to Multiple Imputation by Chained Equations. American Journal of Epidemiology. 2022;191(3):516-525.
}
\keyword{internal}
